Consideration of implementing 11 and 17
The tool needs to quickly ascertain whether the run contained a significant record or scoring highlight. To keep it neat this can occur on right side of each stage in the FINAL STATS section. Worst is probably unnecessary and clutters the tight data set, but may be useful for KT conversion. 

Ultimately, the tool should check against a snapshot or text that contains any and all leaderboards and then makes suggestions to submit and claim the rank/record to encourage broader participation. DKF is a given. MARP and TG make sense in this regard. Going beyond that may make for a difficult task to the (manual) updater, but if this process is AI driven and reviewed then AI can intuitively recognize if the end user has already submitted or needs to and constantly update the text/snapshot leaderboards its using for data.  

Final thought: I don't want to scan down the entire data set checking the pie/spring/rivet stages for the outliers. I just want to know at the bottom and move on from the Inp in most cases.