Generate Log Yes/No after Final Stats? vs. Automatic Output
Along the same train of thought... Quick checker for ISLR score discovery might be down the line, but as I run across archived duplicates I see a use case for the option to halt the log output with a Y/N? in script. Unsure if a universal command would grant permissions to perform the delete log set task in different O/S's folder systems after the logging process, though, this would accomplish the same.